+++ midend/convertErrors.h
@@ -16,98 +16,77 @@ limitations under the License.
 #ifndef _MIDEND_CONVERTERRORS_H_
 #define _MIDEND_CONVERTERRORS_H_
 
+#include <map>
+
+#include "frontends/common/resolveReferences/referenceMap.h"
 #include "frontends/p4/typeChecking/typeChecker.h"
+#include "frontends/p4/typeMap.h"
 #include "ir/ir.h"
+#include "ir/node.h"
+#include "ir/pass_manager.h"
+#include "ir/visitor.h"
+#include "lib/cstring.h"
+#include "lib/null.h"
+#include "lib/safe_vector.h"
+#include "midend/convertEnums.h"
 
 namespace P4 {
 
 /**
- * Policy function: Determines if the error type should be converted to Ser_Enum.
- * This Serializable Enum can later be eliminated by replacing the constant by
- * their respective values.
+ * Policy function: given a number of error values should return
+ * the size of a Type_Bits type used to represent the values.
+ * This class is lefted from enum conversion path.
  */
 class ChooseErrorRepresentation {
  public:
-    virtual ~ChooseErrorRepresentation() {}
+    virtual ~ChooseErrorRepresentation() = default;
     // If true this type has to be converted.
-    virtual bool convert(const IR::Type_Error* /* type */) const { return false; };
+    virtual bool convert(const IR::Type_Error* type) const = 0;
     // errorCount is the number of different error values.
     // The returned value is the width of Type_Bits used
-    // to represent the errors.  Obviously, we must have
+    // to represent the error.  Obviously, we must have
     // 2^(return) >= errorCount.
-    virtual unsigned errorSize(unsigned /* errorCount */) const { return 0; };
-    // This function allows backends to override the values for the error constants.
-    // Default values for error constants is a sequence of numbers starting with 0.
-    virtual IR::IndexedVector<IR::SerEnumMember>* assignValues(IR::Type_Error* type,
-                                                               unsigned width) const {
-        auto members = new IR::IndexedVector<IR::SerEnumMember>;
-        unsigned idx = 0;
-        for (auto d : type->members) {
-            members->push_back(new IR::SerEnumMember(
-                d->name.name, new IR::Constant(IR::Type_Bits::get(width), idx++)));
-        }
-        return members;
-    }
+    virtual unsigned errorSize(unsigned errorCount) const = 0;
 };
 
-/** implement a pass to convert Type_Error to Type_SerEnum
- *
- *  User must provide a class to extend ChooseErrorRepresentation
- *  to specify the width of the generated Type_Bits. User must
- *  also implement a policy to decide whether an Error type
- *  shall be converted.
- *
- *  Example:
- *  error {
- *      ....
- *  }
- *  if the policy is to convert error to serializable enum with underlying type to bit<16>,
- *  then the above is replaced by
- *
- *  enum bit<width> error {
- *      ....
- *  }
- *
- *  @pre none
- *  @post all Type_Error types accepted by 'policy' are converted.
- */
 class DoConvertErrors : public Transform {
     friend class ConvertErrors;
 
+    std::map<cstring, P4::EnumRepresentation*> repr;
     ChooseErrorRepresentation* policy;
+    P4::TypeMap* typeMap;
 
  public:
-    explicit DoConvertErrors(ChooseErrorRepresentation* policy) : policy(policy) {
+    DoConvertErrors(ChooseErrorRepresentation* policy, P4::TypeMap* typeMap)
+        : policy(policy), typeMap(typeMap) {
         CHECK_NULL(policy);
+        CHECK_NULL(typeMap);
         setName("DoConvertErrors");
     }
-
-    const IR::Node* preorder(IR::Type_Error* type) {
-        bool convert = policy->convert(type);
-        if (!convert) return type;
-        IR::IndexedVector<IR::SerEnumMember> members;
-        unsigned count = type->members.size();
-        unsigned width = policy->errorSize(count);
-        return new IR::Type_SerEnum("error", IR::Type_Bits::get(width),
-                                    *policy->assignValues(type, width));
-    }
+    const IR::Node* preorder(IR::Type_Error* type) override;
+    const IR::Node* postorder(IR::Type_Name* type) override;
+    const IR::Node* postorder(IR::Member* member) override;
 };
 
 class ConvertErrors : public PassManager {
     DoConvertErrors* convertErrors{nullptr};
 
  public:
-    ConvertErrors(ReferenceMap* refMap, TypeMap* typeMap, ChooseErrorRepresentation* policy,
-                  TypeChecking* typeChecking = nullptr)
-        : convertErrors(new DoConvertErrors(policy)) {
-        if (!typeChecking) typeChecking = new TypeChecking(refMap, typeMap);
+    using ErrorMapping = decltype(DoConvertErrors::repr);
+    ConvertErrors(P4::ReferenceMap* refMap, P4::TypeMap* typeMap, ChooseErrorRepresentation* policy,
+                  P4::TypeChecking* typeChecking = nullptr)
+        : convertErrors(new DoConvertErrors(policy, typeMap)) {
+        if (typeChecking == nullptr) {
+            typeChecking = new P4::TypeChecking(refMap, typeMap);
+        }
         passes.push_back(typeChecking);
         passes.push_back(convertErrors);
-        passes.push_back(new ClearTypeMap(typeMap));
+        passes.push_back(new P4::ClearTypeMap(typeMap));
         setName("ConvertErrors");
     }
-};
 
+    ErrorMapping getErrorMapping() const { return convertErrors->repr; }
+};
 }  // namespace P4
 
 #endif /* _MIDEND_CONVERTERRORS_H_ */
